changed message handler on the arduino to not use an internal buffer but instead read from the incomming buffer as it arrives. Not sure if this is better or not, will experiment more.
Also changed the escape character to 0x81 instead of 0x55 since the new method doesn't check for 0x55 escaping for the message type byte, so wanted to support up to 128 different message types without worrying about the escape byte. As I'm writing this, I noticed that nothing was checking if the message byte was 0x55 while writing either, so maybe this is not required?

Finally, realized that the escaping buffer in c# was expensive and I really don't care about the exact value for RGB colors, so added logic to optionally convert the value to one lower if its the escape character instead of duplicating the buffer and escaping those values.